% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.r
\name{as.qtDate}
\alias{as.qtDate}
\title{Cast Date to Ctring}
\usage{
as.qtDate(val = Sys.Date())
}
\arguments{
\item{val}{a date string as yyyy-MM-dd}
}
\value{
character
}
\description{
Cast Date to Character
}
\examples{
as.qtDate()
as.qtDate('2015-10-10')
}
\author{
Dan Zhang
}

