% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getCompanyInfo}
\alias{getCompanyInfo}
\title{Get Company Info}
\usage{
getCompanyInfo(data, CompanyCode, date, key)
}
\arguments{
\item{data}{character}

\item{CompanyCode}{character}

\item{date}{character}

\item{key}{character}
}
\description{
Get Company Info
}
\examples{
\dontrun{
getCompanyInfo(data='shareStru',CompanyCode='224448',date='2016-01-07',key=key)
}
}
\author{
Yong Zhou
}

