% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getDaily}
\alias{getDaily}
\title{Get Daily Quote data}
\usage{
getDaily(data, qtid = c(), date = c(), key = NULL)
}
\arguments{
\item{data}{character}

\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get Daily Quote data
}
\author{
Dan Zhang
}

