% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getDailyQuote}
\alias{getDailyQuote}
\title{Get Daily Quote}
\usage{
getDailyQuote(data, qtid = c(), startdate = NULL, enddate = NULL,
  SecuMarket = NULL, ProductCode = c(), key)
}
\arguments{
\item{data}{character}

\item{qtid}{vector}

\item{startdate}{Date}

\item{enddate}{Date}

\item{SecuMarket}{character}

\item{ProductCode}{vector}

\item{key}{character}
}
\description{
Get Daily Quote
}
\examples{
\dontrun{
getDailyQuote(data='mktDaily',qtid=c('000001.SZ','000002.SZ'),key=key)
getDailyQuote(data='mktDaily',startdate='2015-10-01',enddate='2015-10-10',key=key)
getDailyQuote(data='mktFwdDaily',qtid=c('000001.SZ','000002.SZ'),key=key)
getDailyQuote(data='mktFwdDaily',startdate='2015-10-01',enddate='2015-10-10',key=key)
getDailyQuote(data='mktDataIndex',qtid=c('000001.SH','000002.SH'),key=key)
getDailyQuote(data='mktDataIndex',startdate='2015-10-01',enddate='2015-10-10',key=key)
getDailyQuote(data='securitiesMargin',SecuMarket=90,key=key)
getDailyQuote(data='securitiesMargin',startdate='2015-10-01',enddate='2015-10-10',key=key)
}
}
\author{
Dan Zhang
}

