% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getDate}
\alias{getDate}
\title{Get Date date}
\usage{
getDate(data, startdate = NULL, enddate = Sys.Date(), key = NULL)
}
\arguments{
\item{data}{character}

\item{startdate}{Date}

\item{enddate}{Date}

\item{key}{character}
}
\description{
Get Date date
}
\examples{
\dontrun{
getDate(data='tradingDay',key=key)
getDate(data='tradingDay',startdate='2015-10-10',key=key)
getDate(data='tradingDay',startdate='2015-10-10',enddate='2015-12-30',key=key)
}
}
\author{
Dan Zhang
}

