% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getFinancialIndex}
\alias{getFinancialIndex}
\title{Get getFinancialIndex data}
\usage{
getFinancialIndex(qtid = c(), date = c(), key)
}
\arguments{
\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get getFinancialIndex data
}
\author{
Dan Zhang
}

