% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getFwdMktDaily}
\alias{getFwdMktDaily}
\title{Get FwdMktDaily data}
\usage{
getFwdMktDaily(qtid = c(), date = c(), key = NULL)
}
\arguments{
\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get FwdMktDaily data
}
\author{
Dan Zhang
}

