% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getIndustry}
\alias{getIndustry}
\title{Get Industry data}
\usage{
getIndustry(data, date, qtid = c(), CompanyCode = NULL, SW1 = NULL,
  SW2 = NULL, SW3 = NULL, qtind = NULL, key)
}
\arguments{
\item{data}{character}

\item{date}{Date}

\item{qtid}{vector}

\item{CompanyCode}{character}

\item{SW1}{character}

\item{SW2}{character}

\item{SW3}{character}

\item{qtind}{character}

\item{key}{character}
}
\description{
Get Industry data
}
\examples{
\dontrun{
getIndustry(data='industryType',date='2015-12-02',key=key)
getIndustry(data='industryType',date='2015-12-02',qtid=c('000001.SZ','000002.SZ'),key=key)
getIndustry(data='industryType',date='2015-12-02',CompanyCode=6,key=key)
}
}
\author{
Dan Zhang
}

