% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getIndustryType}
\alias{getIndustryType}
\title{Get getIndustryType data}
\usage{
getIndustryType(qtid = c(), date = NULL, CompanyCode = NULL, sw1 = NULL,
  sw2 = NULL, sw3 = NULL, key = NULL)
}
\arguments{
\item{qtid}{vector}

\item{date}{Date}

\item{CompanyCode}{character}

\item{sw1}{character}

\item{sw2}{character}

\item{sw3}{character}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get getIndustryType data
}
\author{
Dan Zhang
}

