% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getKeyMap}
\alias{getKeyMap}
\title{Get getKeyMap data}
\usage{
getKeyMap(qtid = c(), SecuCode = NULL, CompanyCode = NULL,
  ChiName = NULL, SecuMarket = NULL, key)
}
\arguments{
\item{qtid}{character}

\item{SecuCode}{character}

\item{CompanyCode}{character}

\item{ChiName}{character}

\item{SecuMarket}{numeric}

\item{key}{character require}
}
\value{
data.frame
}
\description{
Get getKeyMap data
}
\author{
Dan Zhang
}

