% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getMktDaily}
\alias{getMktDaily}
\title{Get getMktDaily data}
\usage{
getMktDaily(qtid = c(), date = c(), key)
}
\arguments{
\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get getMktDaily data
}
\author{
Dan Zhang
}

