% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getMktDataIndex}
\alias{getMktDataIndex}
\title{Get MktDataIndex data}
\usage{
getMktDataIndex(qtid = c(), date = c(), key)
}
\arguments{
\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get MktDataIndex data
}
\author{
Dan Zhang
}

