% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getQtStock}
\alias{getQtStock}
\title{Get Date date}
\usage{
getQtStock(data, qtid = c(), startdate = NULL, enddate = NULL, key)
}
\arguments{
\item{data}{character}

\item{qtid}{character}

\item{startdate}{Date}

\item{enddate}{Date}

\item{key}{character}
}
\description{
Get Qutke Stock data
}
\examples{
\dontrun{
getQtStock(data='stockBeta',qtid=c('000001.SZ','000002.SZ'),key=key)
getQtStock(data='stockBeta',startdate='2015-10-01',enddate='2015-10-10',key=key)
getQtStock(data='financialIndex',qtid=c('000001.SZ','000002.SZ'),key=key)
getQtStock(data='financialIndex',startdate='2015-10-10',enddate='2015-12-30',key=key)
}
}
\author{
Dan Zhang
}

