% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getSecuritiesMargin}
\alias{getSecuritiesMargin}
\title{Get getSecuritiesMargin data}
\usage{
getSecuritiesMargin(date = c(), SecuMarket = NULL, key)
}
\arguments{
\item{date}{vector}

\item{SecuMarket}{character}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get getSecuritiesMargin data
}
\author{
Dan Zhang
}

