% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getStockBeta}
\alias{getStockBeta}
\title{Get getStockBeta data}
\usage{
getStockBeta(qtid = c(), date = c(), key = NULL)
}
\arguments{
\item{qtid}{vector}

\item{date}{vector}

\item{key}{character}
}
\value{
data.frame
}
\description{
Get getStockBeta data
}
\author{
Dan Zhang
}

