% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getSW1}
\alias{getSW1}
\title{Get Industry FirstIndustryNames}
\usage{
getSW1(industryDF)
}
\arguments{
\item{industryDF}{data.frame}
}
\description{
Get Industry FirstIndustryNames
}
\examples{
\dontrun{
getSW1(industryDF=YourIndustryDataFrame)
}
}
\author{
Yong Zhou
}

