% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getSW2}
\alias{getSW2}
\title{Get Industry SecondIndustryNames}
\usage{
getSW2(industryDF)
}
\arguments{
\item{industryDF}{data.frame}
}
\description{
Get Industry SecondIndustryNames
}
\examples{
\dontrun{
getSW1(industryDF=YourIndustryDataFrame)
}
}
\author{
Yong Zhou
}

