% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getSW3}
\alias{getSW3}
\title{Get Industry ThirdIndustryNames}
\usage{
getSW3(industryDF)
}
\arguments{
\item{industryDF}{data.frame}
}
\description{
Get Industry ThirdIndustryNames
}
\examples{
\dontrun{
getSW1(industryDF=YourIndustryDataFrame)
}
}
\author{
Yong Zhou
}

