% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dao.r
\name{getTradingDay}
\alias{getTradingDay}
\title{Get getTradingDay}
\usage{
getTradingDay(key = NULL)
}
\arguments{
\item{key}{character}
}
\value{
data.frame
}
\description{
Get getTradingDay
}
\author{
Dan Zhang
}

