% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.r
\name{getIndustry}
\alias{getIndustry}
\title{Get Date date}
\usage{
getIndustry(data, qtid = c(), date = NULL, CompanyCode = NULL,
  SW1 = NULL, SW2 = NULL, SW3 = NULL, key)
}
\arguments{
\item{data}{character}

\item{qtid}{vector}

\item{date}{Date}

\item{CompanyCode}{character}

\item{SW1}{character}

\item{SW2}{character}

\item{SW3}{character}

\item{key}{character}
}
\description{
Get Industry date
}
\examples{
\dontrun{
getIndustry(data='industryType',date='2015-12-02',key=key)
getIndustry(data='industryType',qtid=c('000001.SZ','000002.SZ'),key=key)
getIndustry(data='industryType',CompanyCode=6,key=key)
}
}
\author{
Dan Zhang
}

